+++ docs/guide/changelog.md
@@ -1,12 +1,18 @@
 # Changelog
 
+## 1.4.8 (2025-02-24)
+
+- Feature
+  - Input helper message ([#0b19d997](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/0b19d99758e20c3c05cd1d4b60e18a76d44bdd4c?refName=refs%2Fheads%2F2025%2FFEATURE%2FPaulo%2FINPUT_HELPER_MESSAGE) by @jbanares)
+- Fix
+  - Fixed snackbar tone for danger and caution ([#23512c95](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/23512c957be34ab04e640091c2b9e3e98db6b1b8?refName=refs%2Fheads%2F2025%2FFEATURE%2FPaulo%2FINPUT_HELPER_MESSAGE) by @jbanares)
+
 ## 1.4.7 (2025-02-24)
 
 - Feature
   - Textarea ([#1c8100f4](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/1c8100f431890eb773e6faf2cc850bcbd388e896?refName=refs%2Fheads%2F2025%2FFeature%2Fandrea%2Ftextarea) by @adestajo)
 
 ## 1.4.6 (2025-02-24)
-
 - Fix
   - DQA for sidenav and sidepanel ([#e5d81457](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/e5d81457a49c5755782bf0e176d1dbb0645e747e?refName=refs/heads/2025/FIX/ELI/SIDENAV) by @elumilay)